// app/back/src/domain/models/site-funnel.ts
export interface SiteStepsSummary {
  id: number,
  sales: number,
  revenue: number,
  total_step_1: number
  step_1_views: number
  total_step_2: number
  step_2_views: number
  total_step_3: number
  step_3_views: number
  total_step_4: number
  step_4_views: number
  total_checkout: number
  checkout_views: number
}